example_agents/stock_price/requirements.txt
python-dotenv=>1.0.1
pyautogen=>0.3.0
